Add getEncodedValue function
Returns the encoded string representing a Task object to be saved in the storage
