+++ src/main/java/Tasks/ToDo.java
@@ -1,0 +1,30 @@
+package Tasks;
+
+import Tasks.Task;
+import dukeExceptions.DukeException;
+import dukeExceptions.MissingDescriptionException;
+
+/**
+ * Todo represents a Tasks.Task without any datetime objects associated with it.
+ * Inherits from Tasks.Task.
+ */
+public class ToDo extends Task {
+    public ToDo(String taskDescription) {
+        super(taskDescription);
+    }
+
+    public ToDo(String taskDescription, boolean isComplete) {
+        super(taskDescription);
+        setIsComplete(isComplete);
+    }
+
+    @Override
+    public String getEncodedValue() {
+        return String.format("[T]#%s#%s", getTaskDescription(), getIsComplete());
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
